refactor(DP-870): decompose RuleAgeSelector into focused sub-components (#413)
Split the monolithic RuleAgeSelector (420 lines) into single-responsibility
files: RuleAgeSelectorSingleBound (Bunny/SingleBoundSelector path with named
handleConstraintChange), RuleAgeSelectorSlider (legacy slider path with named
handleSliderChange), RuleAgeSelectorReadOnly, and AgeInput. RuleAgeSelector
becomes a thin ~55-line dispatcher. RuleAgeFilter updated to render
RuleAgeSelectorReadOnly directly, bypassing the full interactive component.
Adds tests covering feature flag branching and readOnly rendering.
