+++ src/components/RuleAgeSelector/index.ts
@@ -1,3 +1,4 @@
 import RuleAgeSelector from "./RuleAgeSelector";
 
+export { default as RuleAgeSelectorReadOnly } from "./RuleAgeSelectorReadOnly";
 export default RuleAgeSelector;
